Confirmed Command3 is queue a unit.
There does not appear to be any even later on when the unit is spawned.
This suggests it replies on the game to simulate the gamestate forward
in a deterministic way (so when later commands refer to the unit by ID
the ID is valid because it was deterministically decided).